backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "bcrypt>=5.0.0",
    "bs4>=0.0.2",
    "flask>=3.1.2",
    "flask-cors>=6.0.2",
    "flask-limiter>=4.1.1",
    "langchain>=0.3.0",
    "langchain-openai>=0.3.0",
    "langgraph>=0.4.0",
    "pillow>=11.0.0",
    "PyJWT>=2.8.0",
    "psycopg[binary]>=3.3.2",
    "redis>=7.1.0",
    "requests>=2.32.5",
]
